requirements.txt
# ─── Core Web Framework ───────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── OCR & Image Processing ───────────────────────────────────────────────────
pytesseract==0.3.10
pdf2image==1.17.0
opencv-python==4.9.0.80
Pillow==10.3.0

# ─── LLM & LangChain ──────────────────────────────────────────────────────────
groq>=0.9.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
openai>=1.30.0
langchain-core>=0.2.0

# ─── Embeddings & Vector Store ────────────────────────────────────────────────
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# ─── Data Validation ──────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Frontend ─────────────────────────────────────────────────────────────────
streamlit==1.33.0
requests==2.31.0
plotly>=5.20.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
numpy==1.26.4
tiktoken==0.7.0
pypdf==4.2.0
aiofiles==23.2.1
httpx==0.27.0

# ─── Enterprise Features (new) ────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openpyxl>=3.1.0
networkx>=3.3
